+++ common/src/interface.ts
@@ -1,0 +1,9 @@
+export interface CommonMeta {
+  parameters: {
+    [key: string]: number[];
+  };
+}
+
+export interface CommonClass extends Object {
+  __meta__: CommonMeta;
+}